ikram berradi

+++ ma.atos.bankmanagement/src/main/java/ma/atos/ma/atos/bankmanagement/services/TierServiceImpl.java
@@ -6,6 +6,7 @@ import ma.atos.ma.atos.bankmanagement.entities.PersonneMorale;
 import ma.atos.ma.atos.bankmanagement.entities.PersonnePhysique;
 
 import ma.atos.ma.atos.bankmanagement.entities.Tier;
+import ma.atos.ma.atos.bankmanagement.exceptions.CompteException;
 import ma.atos.ma.atos.bankmanagement.exceptions.TierNotFoundExeption;
 import ma.atos.ma.atos.bankmanagement.mappers.PersonneMoraleMapper;
 import ma.atos.ma.atos.bankmanagement.mappers.PersonnePhysiqueMapper;
@@ -28,7 +29,6 @@ public class TierServiceImpl implements TierService {
     PersonnePhysiqueMapper personnePhysiqueMapper;
 @Autowired
    TierRepository tierRepository;
-
   /*  @Override
     public List<PersonneMoraleDto> getAllPersonneMorale() {
         List<Tier> personneMorales= tierRepository.findByTierType("PM");
@@ -36,7 +36,6 @@ public class TierServiceImpl implements TierService {
                 personneMoraleMapper.personneToPersonneDto((PersonneMorale) personneMorale)).collect(Collectors.toList());
         return personneMoraleDtos;
     }*/
-
     @Override
     public List<PersonneMoraleDto> getAllPersonneMorale() {
         List<PersonneMorale> personneMorales= tierRepository.findPMByTierType("PM");
@@ -48,7 +47,7 @@ public class TierServiceImpl implements TierService {
     @Override
     public PersonneMoraleDto getPersonneMorale(Long id) throws TierNotFoundExeption {
         PersonneMorale pp = (PersonneMorale) tierRepository.findByTierTypeEqualsAndIdClientEquals("PM", id);
-        log.info(pp.toString());
+        log.info("ID du tiers est " + id);
         PersonneMorale personneMorale = (PersonneMorale) tierRepository.findById(id).orElseThrow(() ->
                 new TierNotFoundExeption(messageSource.getMessage("Tier.not.found.exception",new Object[]{}, Locale.getDefault())));
         return personneMoraleMapper.personneToPersonneDto(personneMorale);
@@ -61,10 +60,8 @@ public class TierServiceImpl implements TierService {
     }
     @Override
     public void deletPersonneMorale(long id) {
-
         tierRepository.deleteById(id);
     }
-
     @Override
     public List<PersonnePhysiqueDto> getAllPersonnePhysique() {
         List<PersonnePhysique> personnePhysiques= tierRepository.findPPByTierType("PP");
@@ -74,28 +71,30 @@ public class TierServiceImpl implements TierService {
         });
         return personnePhysiqueDtos;
     }
-
     @Override
     public PersonnePhysiqueDto getPersonnePhysique(Long id) throws TierNotFoundExeption {
         PersonnePhysique pp = (PersonnePhysique) tierRepository.findByTierTypeEqualsAndIdClientEquals("PP", id);
-        log.info(pp.toString());
+        log.info("ID du tiers est " + id);
         PersonnePhysique personnePhysique = (PersonnePhysique) tierRepository.findById(id).orElseThrow(() ->
-                new TierNotFoundExeption(messageSource.getMessage("Tier Not Found Exeption",new Object[]{}, Locale.getDefault())));
+                new TierNotFoundExeption("Tier not found ","Tier.not.found.exception"));
         return personnePhysiqueMapper.personnePhysiqueToPersonnePhysiqueDto(personnePhysique);
-
     }
-
     @Override
     public PersonnePhysiqueDto savePersonnePhysique(PersonnePhysiqueDto personnePhysiqueDto) {
         PersonnePhysique personnePhysique =personnePhysiqueMapper.PpDtoToPp(personnePhysiqueDto);
         PersonnePhysique savePersonnePhysique=  tierRepository.save(personnePhysique);
         return personnePhysiqueMapper.personnePhysiqueToPersonnePhysiqueDto(savePersonnePhysique);
     }
-
     @Override
-    public void deletPersonnePhysique(long id) {
-        tierRepository.deleteById(id);
+    public void deletPersonnePhysique(long id) throws TierNotFoundExeption {
+        if (tierRepository.findById(id) ==null) {
+
+            throw new TierNotFoundExeption("Tier not found ","Tier.not.found.exception");
+        } else {
+            tierRepository.deleteById(id);
+        }
     }
+
   /*  @Override
     public TierDto getTier(long Id) throws TierNotFoundExeption {
